Add code for user input Type and Range checking
This removed a bug that allowed users to enter any value for the start
and end coordinates which made the start_x, start_y, end_x, and end_y
to be initialized to 0 respectively if any of them got an invalid type.
The code also removed a bug that coudl allow a user to enter values that
are greater than 100 or less than 0.